learning 2-agent cartpole with shared, global observation, single policy (outputting multi-discrete actions), and separate reward streams and value heads.
